# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QAV API
  version: 1.0.0
  description: Quantum-resistant Access Vault API - Secure encrypted file storage and sharing with end-to-end encryption
  contact:
    name: QAV Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.qav.io
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoints
  schemas:
...
    SRPInitRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    SRPInitResponse:
      type: object
      properties:
        salt:
          type: string
          description: Base64-encoded salt for client-side hashing
          example: "aBc1De2Fg3Hi4Jk5Lm6No7Pq8Rs9Tu0Vw1Xy2Za3="
        B:
          type: string
          description: Server's ephemeral value for SRP
          example: "fEhKJ7vKL8mN9oP0qR1sT2uV3wX4yZ5aB6cD7eF8gH9iJ0kL1mN2oP3qR4sTu5vW6xY7zA8bC9dE0fG1hI2jK3lM4nO5pQ6rS7tU8vW9xY0zAaA=="
        session_id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
...
    ValidationError:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
        fields:
          type: object
...
paths:
...
  /api/v1/auth/srp/init:
    post:
      tags:
        - Auth
      summary: Initialize SRP authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SRPInitRequest'
      responses:
        '200':
          description: SRP session initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SRPInitResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          description: Rate limited
        '500':
          description: Internal server error
  /api/v1/auth/srp/verify:
    post:
      tags:
        - Auth
      summary: Verify SRP authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SRPVerifyRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SRPVerifyResponse'
        '400':
          description: Invalid SRP proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
        '429':
          description: Rate limited or account locked
        '500':
          description: Internal server error
  /api/v1/auth/fido2/challenge:
    post:
      tags:
        - Auth
      summary: Get WebAuthn challenge for passwordless login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FIDO2ChallengeRequest'
      responses:
        '200':
          description: Challenge generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FIDO2ChallengeResponse'
        '404':
          description: User not found
        '400':
          description: User has no FIDO2 credentials registered
        '500':
          description: Internal server error
  /api/v1/auth/fido2/verify:
    post:
      tags:
        - Auth
      summary: Verify WebAuthn assertion for login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FIDO2VerifyRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FIDO2VerifyResponse'
        '400':
          description: Invalid assertion
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          description: Invalid or expired refresh token
        '500':
          description: Internal server error
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout and revoke tokens
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "logged_out"
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # ====== Vaults ======
  /api/v1/vaults:
    post:
      tags:
        - Vaults
      summary: Create a new vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVaultRequest'
      responses:
        '201':
          description: Vault created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVaultResponse'
        '400':
          description: Invalid metadata
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    get:
      tags:
        - Vaults
      summary: List user's vaults
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of vaults
          content:
            application/json:
              schema:
                type: object
                properties:
                  vaults:
                    type: array
                    items:
                      $ref: '#/components/schemas/VaultSummary'
                  total_count:
                    type: integer
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/vaults/{vaultID}:
    get:
      tags:
        - Vaults
      summary: Get vault details
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vault details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultSummary'
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault not found
        '500':
          description: Internal server error

    put:
      tags:
        - Vaults
      summary: Update vault metadata
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVaultRequest'
      responses:
        '200':
          description: Vault updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "updated"
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault not found
        '500':
          description: Internal server error

    delete:
      tags:
        - Vaults
      summary: Delete vault (soft delete)
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Vault deleted
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault not found
        '500':
          description: Internal server error

  # ====== Key Hierarchy ======
  /api/v1/vaults/{vaultID}/key-hierarchy:
    post:
      tags:
        - Vaults
      summary: Store vault key hierarchy (wrapped keys)
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyHierarchyRequest'
      responses:
        '201':
          description: Key hierarchy stored
        '400':
          description: Invalid key data
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault not found
        '500':
          description: Internal server error

    get:
      tags:
        - Vaults
      summary: Retrieve vault key hierarchy
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Key hierarchy data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyHierarchyResponse'
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault or keys not found
        '500':
          description: Internal server error

  # ====== Files/Blobs ======
...
  /api/v1/vaults/{vaultID}/rotation-status:
    get:
      tags:
        - Vaults
      summary: Get vault key rotation status
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rotation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyRotationStatus'
        '401':
          description: Unauthorized
        '403':
          description: Access denied
        '404':
          description: Vault not found
        '500':
          description: Internal server error

  # ====== Sharing ======
...
  /api/v1/user/account:
    delete:
      tags:
        - User
      summary: Delete user account (irreversible)
      responses:
        '204':
          description: Account deleted
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  # ====== Sync ======
...